+++ packages/jest-reporters/src/__tests__/NotifyReporter.test.ts
@@ -224,14 +224,13 @@ describe('node-notifier is an optional dependency', () => {
 
   test('without node-notifier uses mock function that throws an error', () => {
     jest.doMock('node-notifier', () => {
-      const error: unknown = new Resolver.ModuleNotFoundError(
+      throw new Resolver.ModuleNotFoundError(
         "Cannot find module 'node-notifier'",
       );
-      throw error;
     });
 
     expect(ctor).toThrow(
-      'notify reporter requires optional dependeny node-notifier but it was not found',
+      'notify reporter requires optional peer dependency node-notifier but it was not found',
     );
   });